split types
