+++ src/components/word/KanjiDetailsModal.tsx
@@ -2,9 +2,10 @@
 
 import useModal from "@/hooks/useModal";
 import BaseModal from "@/components/ui/BaseModal";
-import type { $Kanji, MKanji } from "@/types";
 import KanjiCard from "@/components/kanji/KanjiCard";
 import { useLocale } from "next-intl";
+import type { $Kanji } from "@/types/models";
+import type { MKanji } from "@/types/modals";
 
 const KanjiDetailsModal = () => {
   const { modalOptions } = useModal();
