+++ src/components/word/WordCard.tsx
@@ -1,9 +1,10 @@
-import type { $Word, PropsWithLocalization } from "@/types";
 import { Box, Card, CardContent } from "@mui/material";
 import Table, { Row } from "@/components/ui/Table";
 import { ClickableKanji } from "./ClickableKanji";
 import getLocalizedTranslations from "@/lib/utils/getLocalizedTranslations";
 import extractKanjiFromWord from "@/lib/utils/extractKanjiFromWord";
+import type { PropsWithLocalization } from "@/types/utils";
+import type { $Word } from "@/types/models";
 
 const WordCard = ({
   value,
