+++ src/types/form.ts
@@ -1,0 +1,6 @@
+import type { ApiResponseStatus } from "./api";
+
+export type FormState<T> = Partial<{
+  apiResponse: ApiResponseStatus;
+  validationErrors: Record<keyof T, string>;
+}>;
