+++ src/types/modals.ts
@@ -1,0 +1,3 @@
+export type MKanji = "add-kanji" | "kanji-details";
+
+export type MWord = "add-word";
